Data.Text.Internal.Builder.RealFloat.Functions

Description

Warning: this is an internal module, and does not have a stable API or name. Functions in this module may not check or enforce preconditions expected by public modules. Use at your own risk!

Documentation

roundTo :: Int -> [Int] -> (Int, [Int])