Data.Text.Internal.Encoding.Utf16

Description

Warning: this is an internal module, and does not have a stable API or name. Functions in this module may not check or enforce preconditions expected by public modules. Use at your own risk!

Basic UTF-16 validation and character manipulation.

Documentation

chr2 :: Word16 -> Word16 -> Char

validate1 :: Word16 -> Bool

validate2 :: Word16 -> Word16 -> Bool