Data.Text.Internal.Encoding.Utf8

Description

Warning: this is an internal module, and does not have a stable API or name. Functions in this module may not check or enforce preconditions expected by public modules. Use at your own risk!

Basic UTF-8 validation and character manipulation.

Documentation

utf8Length :: Char -> Int

Since: text-2.0

utf8LengthByLeader :: Word8 -> Int

Since: text-2.0

ord2 :: Char -> (Word8, Word8)

ord3 :: Char -> (Word8, Word8, Word8)

ord4 :: Char -> (Word8, Word8, Word8, Word8)

chr2 :: Word8 -> Word8 -> Char

chr3 :: Word8 -> Word8 -> Word8 -> Char

chr4 :: Word8 -> Word8 -> Word8 -> Word8 -> Char

Validation

validate1 :: Word8 -> Bool

validate2 :: Word8 -> Word8 -> Bool

validate3 :: Word8 -> Word8 -> Word8 -> Bool

validate4 :: Word8 -> Word8 -> Word8 -> Word8 -> Bool

Naive decoding

newtype DecoderState

Constructors

DecoderState Word8

utf8AcceptState :: DecoderState

utf8RejectState :: DecoderState

updateDecoderState :: Word8 -> DecoderState -> DecoderState

data DecoderResult

Since: text-2.0

Constructors

Accept !Char

Incomplete !DecoderState !CodePoint

Reject

newtype CodePoint

Constructors

CodePoint Int

utf8DecodeStart :: Word8 -> DecoderResult

Since: text-2.0

utf8DecodeContinue :: Word8 -> DecoderState -> CodePoint -> DecoderResult

Since: text-2.0