Data.Text.Internal.Functions

Description

Warning: this is an internal module, and does not have a stable API or name. Functions in this module may not check or enforce preconditions expected by public modules. Use at your own risk!

Useful functions.

Documentation

intersperse :: a -> [a] -> [a]

A lazier version of Data.List.intersperse. The other version causes space leaks!

unsafeWithForeignPtr :: ForeignPtr a -> (Ptr a -> IO b) -> IO b