Data.Text.Internal.StrictBuilder

Description

Warning: this is an internal module, and does not have a stable API or name. Functions in this module may not check or enforce preconditions expected by public modules. Use at your own risk!

Since: text-2.0.2

Documentation

data StrictBuilder

A delayed representation of strict Text.

Since: text-2.0.2

Constructors

StrictBuilder

Fields

sbLength :: !Int

sbWrite :: forall s. MArray s -> Int -> ST s ()

toText :: StrictBuilder -> Text

Use StrictBuilder to build Text.

Since: text-2.0.2

fromChar :: Char -> StrictBuilder

Since: text-2.0.2

fromText :: Text -> StrictBuilder

Copy Text in a StrictBuilder

Since: text-2.0.2

Unsafe

For internal purposes, we abuse StrictBuilder as a delayed Array rather than Text: it may not actually be valid Text.

unsafeFromByteString :: ByteString -> StrictBuilder

Copy a ByteString.

Unsafe: This may not be valid UTF-8 text.

Since: text-2.0.2

unsafeFromWord8 :: Word8 -> StrictBuilder

Unsafe: This may not be valid UTF-8 text.

Since: text-2.0.2