Data.Text.Internal.Unsafe.Char

Description

Warning: this is an internal module, and does not have a stable API or name. Functions in this module may not check or enforce preconditions expected by public modules. Use at your own risk!

Fast character manipulation functions.

Documentation

ord :: Char -> Int

unsafeChr16 :: Word16 -> Char

Since: text-2.0

unsafeChr8 :: Word8 -> Char

unsafeChr32 :: Word32 -> Char

unsafeWrite :: MArray s -> Int -> Char -> ST s Int

Write a character into the array at the given offset. Returns the number of Word8s written.