Data.Text.Lazy.Builder.RealFloat

Description

Write a floating point value to a Builder.

Documentation

data FPFormat

Control the rendering of floating point numbers.

Constructors

Exponent
Scientific notation (e.g. 2.3e123).

Fixed
Standard decimal notation.

Generic
Use decimal notation for values between 0.1 and 9,999,999, and scientific notation otherwise.

realFloat :: RealFloat a => a -> Builder

Show a signed RealFloat value to full precision, using standard decimal notation for arguments whose absolute value lies between 0.1 and 9,999,999, and scientific notation otherwise.

formatRealFloat

Arguments

:: RealFloat a

=> FPFormat

-> Maybe Int
Number of decimal places to render.

-> a

-> Builder

Encode a signed RealFloat according to FPFormat and optionally requested precision.

This corresponds to the show{E,F,G}Float operations provided by base's Numeric module.

NOTE: The functions in base-4.12 changed the serialisation in case of a Just 0 precision; this version of text still provides the serialisation as implemented in base-4.11. The next major version of text will switch to the more correct base-4.12 serialisation.