Data.Text.Internal.Builder.Functions

Description

Warning: this is an internal module, and does not have a stable API or name. Functions in this module may not check or enforce preconditions expected by public modules. Use at your own risk!

Useful functions and combinators.

Documentation

(<>) :: Builder -> Builder -> Builder

The normal mappend function with right associativity instead of left.

i2d :: Int -> Char

Unsafe conversion for decimal digits.