Data.Text.Internal.Encoding.Fusion

Description

Warning: this is an internal module, and does not have a stable API or name. Functions in this module may not check or enforce preconditions expected by public modules. Use at your own risk!

Fusible Stream-oriented functions for converting between Text and several common encodings.

Streaming

streamASCII :: ByteString -> Stream Char

Deprecated: Do not use this function

streamUtf8 :: OnDecodeError -> ByteString -> Stream Char

O(n) Convert a ByteString into a 'Stream Char', using UTF-8 encoding.

streamUtf16LE :: OnDecodeError -> ByteString -> Stream Char

O(n) Convert a ByteString into a 'Stream Char', using little endian UTF-16 encoding.

streamUtf16BE :: OnDecodeError -> ByteString -> Stream Char

O(n) Convert a ByteString into a 'Stream Char', using big endian UTF-16 encoding.

streamUtf32LE :: OnDecodeError -> ByteString -> Stream Char

O(n) Convert a ByteString into a 'Stream Char', using little endian UTF-32 encoding.

streamUtf32BE :: OnDecodeError -> ByteString -> Stream Char

O(n) Convert a ByteString into a 'Stream Char', using big endian UTF-32 encoding.

Unstreaming

unstream :: Stream Word8 -> ByteString

O(n) Convert a Stream Word8 to a ByteString.

module Data.Text.Internal.Encoding.Fusion.Common