Data.Text.Internal.Encoding.Fusion.Common

Description

Warning: this is an internal module, and does not have a stable API or name. Use at your own risk!

Fusible Stream-oriented functions for converting between Text and several common encodings.

Restreaming

restreamUtf16LE :: Stream Char -> Stream Word8

restreamUtf16BE :: Stream Char -> Stream Word8

restreamUtf32LE :: Stream Char -> Stream Word8

restreamUtf32BE :: Stream Char -> Stream Word8