Data.Text.Internal.Encoding.Utf32

Description

Warning: this is an internal module, and does not have a stable API or name. Functions in this module may not check or enforce preconditions expected by public modules. Use at your own risk!

Basic UTF-32 validation.

Documentation

validate :: Word32 -> Bool