Data.Text.Internal.Encoding.Utf8

Description

Warning: this is an internal module, and does not have a stable API or name. Functions in this module may not check or enforce preconditions expected by public modules. Use at your own risk!

Basic UTF-8 validation and character manipulation.

Documentation

ord2 :: Char -> (Word8, Word8)

ord3 :: Char -> (Word8, Word8, Word8)

ord4 :: Char -> (Word8, Word8, Word8, Word8)

chr2 :: Word8 -> Word8 -> Char

chr3 :: Word8 -> Word8 -> Word8 -> Char

chr4 :: Word8 -> Word8 -> Word8 -> Word8 -> Char

Validation

validate1 :: Word8 -> Bool

validate2 :: Word8 -> Word8 -> Bool

validate3 :: Word8 -> Word8 -> Word8 -> Bool

validate4 :: Word8 -> Word8 -> Word8 -> Word8 -> Bool