Data.Text.Internal.Fusion

Description

Warning: this is an internal module, and does not have a stable API or name. Functions in this module may not check or enforce preconditions expected by public modules. Use at your own risk!

Text manipulation functions represented as fusible operations over streams.

Types

data Stream a

Constructors

forall s . Stream (s -> Step s a) !s !Size

data Step s a

Intermediate result in a processing pipeline.

Constructors

Done

Skip !s

Yield !a !s

Creation and elimination

stream :: Text -> Stream Char

O(n) Convert a Text into a 'Stream Char'.

unstream :: Stream Char -> Text

O(n) Convert a 'Stream Char' into a Text.

reverseStream :: Text -> Stream Char

O(n) Convert a Text into a 'Stream Char', but iterate backwards.

length :: Stream Char -> Int

Transformations

reverse :: Stream Char -> Text

O(n) Reverse the characters of a string.

Construction

Scans

reverseScanr :: (Char -> Char -> Char) -> Char -> Stream Char -> Stream Char

O(n) Perform the equivalent of scanr over a list, only with the input and result reversed.

Accumulating maps

mapAccumL :: (a -> Char -> (a, Char)) -> a -> Stream Char -> (a, Text)

O(n) Like a combination of map and foldl'. Applies a function to each element of a Text, passing an accumulating parameter from left to right, and returns a final Text.

Generation and unfolding

unfoldrN :: Int -> (a -> Maybe (Char, a)) -> a -> Stream Char

O(n) Like unfoldr, unfoldrN builds a stream from a seed value. However, the length of the result is limited by the first argument to unfoldrN. This function is more efficient than unfoldr when the length of the result is known.

Indexing

index :: Stream Char -> Int -> Char

O(n) stream index (subscript) operator, starting from 0.

findIndex :: (Char -> Bool) -> Stream Char -> Maybe Int

The findIndex function takes a predicate and a stream and returns the index of the first element in the stream satisfying the predicate.

countChar :: Char -> Stream Char -> Int

O(n) The count function returns the number of times the query element appears in the given stream.