Data.Text.Internal.Fusion.Size

Description

Warning: this is an internal module, and does not have a stable API or name. Functions in this module may not check or enforce preconditions expected by public modules. Use at your own risk!

Size hints.

Documentation

data Size

exactly :: Size -> Maybe Int

exactSize :: Int -> Size

maxSize :: Int -> Size

betweenSize :: Int -> Int -> Size

unknownSize :: Size

smaller :: Size -> Size -> Size

Minimum of two size hints.

larger :: Size -> Size -> Size

Maximum of two size hints.

upperBound :: Int -> Size -> Int

Compute the maximum size from a size hint, if possible.

lowerBound :: Int -> Size -> Int

Compute the maximum size from a size hint, if possible.

compareSize :: Size -> Int -> Maybe Ordering

isEmpty :: Size -> Bool