Data.Text.Internal.Fusion.Size

Description

Warning: this is an internal module, and does not have a stable API or name. Functions in this module may not check or enforce preconditions expected by public modules. Use at your own risk!

Size hints.

Documentation

data Size

A size in UTF-16 code units.

Sizes

exactSize :: Int -> Size

maxSize :: Int -> Size

betweenSize :: Int -> Int -> Size

unknownSize :: Size

unionSize :: Size -> Size -> Size

charSize :: Char -> Size

The Size of the given code point.

codePointsSize :: Int -> Size

The Size of n code points.

Querying sizes

exactly :: Size -> Maybe Int

smaller :: Size -> Size -> Size

Minimum of two size hints.

larger :: Size -> Size -> Size

Maximum of two size hints.

upperBound :: Int -> Size -> Int

Compute the maximum size from a size hint, if possible.

lowerBound :: Int -> Size -> Int

Compute the maximum size from a size hint, if possible.

compareSize :: Size -> Size -> Maybe Ordering

Determine the ordering relationship between two Sizes, or Nothing in the indeterminate case.

isEmpty :: Size -> Bool