Data.Text.Internal.Fusion.Types

Description

Warning: this is an internal module, and does not have a stable API or name. Functions in this module may not check or enforce preconditions expected by public modules. Use at your own risk!

Core stream fusion functionality for text.

Documentation

data CC s

Specialised tuple for case conversion.

Constructors

CC !s !Char !Char

data PairS a b

Strict pair.

Constructors

!a :*: !b infixl 2

data Scan s

An intermediate result in a scan.

Constructors

Scan1 !Char !s

Scan2 !Char !s

data RS s

Restreaming state.

Constructors

RS0 !s

RS1 !s !Word8

RS2 !s !Word8 !Word8

RS3 !s !Word8 !Word8 !Word8

data Step s a

Intermediate result in a processing pipeline.

Constructors

Done

Skip !s

Yield !a !s

data Stream a

Constructors

forall s . Stream (s -> Step s a) !s !Size

empty :: Stream a

The empty stream.