Data.Text.Lazy.Builder.RealFloat

Description

Write a floating point value to a Builder.

Documentation

data FPFormat

Control the rendering of floating point numbers.

Constructors

Exponent
Scientific notation (e.g. 2.3e123).

Fixed
Standard decimal notation.

Generic
Use decimal notation for values between 0.1 and 9,999,999, and scientific notation otherwise.

realFloat :: RealFloat a => a -> Builder

Show a signed RealFloat value to full precision, using standard decimal notation for arguments whose absolute value lies between 0.1 and 9,999,999, and scientific notation otherwise.

formatRealFloat

Arguments

:: RealFloat a

=> FPFormat

-> Maybe Int
Number of decimal places to render.

-> a

-> Builder